moving back to mixin approach
