+++ tests/kafkatest/services/console_consumer.py
@@ -19,7 +19,7 @@ import os
 from ducktape.services.background_thread import BackgroundThreadService
 
 from kafkatest.directory_layout.kafka_path import KafkaPathResolverMixin
-from kafkatest.services.monitor.jmx import JmxTool
+from kafkatest.services.monitor.jmx import JmxMixin
 from kafkatest.version import DEV_BRANCH, LATEST_0_8_2, LATEST_0_9, LATEST_0_10_0, V_0_9_0_0, V_0_10_0_0, V_0_11_0_0, V_2_0_0
 
 """
@@ -27,7 +27,7 @@ The console consumer is a tool that reads data from Kafka and outputs it to stan
 """
 
 
-class ConsoleConsumer(KafkaPathResolverMixin, BackgroundThreadService):
+class ConsoleConsumer(KafkaPathResolverMixin, JmxMixin, BackgroundThreadService):
     # Root directory for persistent output
     PERSISTENT_ROOT = "/mnt/console_consumer"
     STDOUT_CAPTURE = os.path.join(PERSISTENT_ROOT, "console_consumer.stdout")
@@ -88,6 +88,8 @@ class ConsoleConsumer(KafkaPathResolverMixin, BackgroundThreadService):
             client_prop_file_override   Override client.properties file used by the consumer
             consumer_properties         A dict of values to pass in as --consumer-property key=value
         """
+        JmxMixin.__init__(self, num_nodes=num_nodes, jmx_object_names=jmx_object_names, jmx_attributes=(jmx_attributes or []),
+                          root=ConsoleConsumer.PERSISTENT_ROOT)
         BackgroundThreadService.__init__(self, context, num_nodes)
         self.kafka = kafka
         self.new_consumer = new_consumer
@@ -99,10 +101,6 @@ class ConsoleConsumer(KafkaPathResolverMixin, BackgroundThreadService):
         self.consumer_timeout_ms = consumer_timeout_ms
         for node in self.nodes:
             node.version = version
-
-        self.jmx_tool = JmxTool(context, num_nodes=num_nodes,
-                                jmx_object_names=jmx_object_names, jmx_attributes=(jmx_attributes or []),
-                                root=ConsoleConsumer.PERSISTENT_ROOT)
 
         self.from_beginning = from_beginning
         self.message_validator = message_validator
@@ -154,7 +152,7 @@ class ConsoleConsumer(KafkaPathResolverMixin, BackgroundThreadService):
         args['log4j_config'] = ConsoleConsumer.LOG4J_CONFIG
         args['config_file'] = ConsoleConsumer.CONFIG_FILE
         args['stdout'] = ConsoleConsumer.STDOUT_CAPTURE
-        args['jmx_port'] = self.jmx_tool.jmx_port
+        args['jmx_port'] = self.jmx_port
         args['console_consumer'] = self.path.script("kafka-console-consumer.sh", node)
         args['broker_list'] = self.kafka.bootstrap_servers(self.security_config.security_protocol)
 
@@ -251,8 +249,8 @@ class ConsoleConsumer(KafkaPathResolverMixin, BackgroundThreadService):
         consumer_output = node.account.ssh_capture(cmd, allow_fail=False)
 
         with self.lock:
-            self.logger.debug("collecting following jmx objects: %s", self.jmx_tool.jmx_object_names)
-            self.jmx_tool.start_jmx_tool(idx, node)
+            self.logger.debug("collecting following jmx objects: %s", self.jmx_object_names)
+            self.start_jmx_tool(idx, node)
 
         for line in consumer_output:
             msg = line.strip()
@@ -268,7 +266,7 @@ class ConsoleConsumer(KafkaPathResolverMixin, BackgroundThreadService):
                     self.messages_consumed[idx].append(msg)
 
         with self.lock:
-            self.jmx_tool.read_jmx_output(idx, node)
+            self.read_jmx_output(idx, node)
 
     def start_node(self, node):
         BackgroundThreadService.start_node(self, node)
@@ -286,7 +284,7 @@ class ConsoleConsumer(KafkaPathResolverMixin, BackgroundThreadService):
         if self.alive(node):
             self.logger.warn("%s %s was still alive at cleanup time. Killing forcefully..." %
                              (self.__class__.__name__, node.account))
-        self.jmx_tool.clean_node(self.idx(node), node)
+        JmxMixin.clean_node(self, node)
         node.account.kill_java_processes(self.java_class_name(), clean_shutdown=False, allow_fail=True)
         node.account.ssh("rm -rf %s" % ConsoleConsumer.PERSISTENT_ROOT, allow_fail=False)
         self.security_config.clean_node(node)
